# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: hybrid-data-svc REST API
  version: 0.1.0
  description: |
    REST gateway over the internal `BarService` gRPC + the curated `assets`
    catalog. Designed to feed Ghostfolio's MANUAL data source and any other
    HTTP-only consumer.

    All `/v1/*` endpoints accept an optional `Authorization: Bearer <token>`
    header; auth is enforced only when the `REST_AUTH_TOKEN` env var is set.
    `/healthz` is always open.

    Symbols use TradingView chart identifiers, uppercase, `:`-separated
    (e.g. `BINANCE:BTCUSDT`, `NASDAQ:AAPL`). URL-encode `:` as `%3A` in
    path parameters.
servers:
  - url: http://localhost:8001
    description: Local docker-compose default.
...
paths:
...
  /v1/historical/{symbol}:
    get:
      tags: [historical]
      summary: OHLCV bars over a date range.
      operationId: getHistorical
      parameters:
        - $ref: "#/components/parameters/SymbolPath"
        - name: from
          in: query
          required: true
          description: Start of the range, inclusive (epoch seconds, UTC).
          schema:
            type: integer
            minimum: 0
        - name: to
          in: query
          required: true
          description: End of the range, inclusive (epoch seconds, UTC).
          schema:
            type: integer
            minimum: 0
        - name: interval
          in: query
          required: false
          description: Bar timeframe.
          schema:
            $ref: "#/components/schemas/Timeframe"
      responses:
        "200":
          description: Bars in the range. Capped at 5000; if more would fit, `truncated` is true and the most recent 5000 are returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoricalResponse"
        "400":
          $ref: "#/components/responses/BadRange"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/UnknownSymbol"
        "422":
          $ref: "#/components/responses/ValidationFailed"
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: |
        Optional bearer auth. When the `REST_AUTH_TOKEN` env var is set,
        every `/v1/*` call must include `Authorization: Bearer <token>`;
        missing/invalid → 401. `/healthz` is always open.
  parameters:
    SymbolPath:
      name: symbol
      in: path
      required: true
      description: TradingView identifier, uppercase, `:`-separated. URL-encode `:` as `%3A`.
      schema:
        type: string
        minLength: 3
        maxLength: 32
        pattern: "^[A-Z0-9]+:[A-Z0-9._/-]+$"
      examples:
        crypto:
          value: BINANCE:BTCUSDT
        equity:
          value: NASDAQ:AAPL
  responses:
    Unauthorized:
      description: Bearer token missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnknownSymbol:
      description: Symbol is not in the asset catalog.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    BadRange:
      description: Invalid query parameters (e.g. `from >= to`, unsupported interval).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationFailed:
      description: Request validation failed (FastAPI default).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPValidationError"
  schemas:
    Timeframe:
      type: string
      enum:
        - "1m"
        - "3m"
        - "5m"
        - "15m"
        - "30m"
        - "1h"
        - "2h"
        - "4h"
        - "1D"
        - "1W"
      default: "1D"
...
    Bar:
      type: object
      required: [ts, open, high, low, close, volume]
      properties:
        ts:
          type: integer
          description: Bar open timestamp, epoch seconds UTC.
        open:
          type: number
          format: double
        high:
          type: number
          format: double
        low:
          type: number
          format: double
        close:
          type: number
          format: double
        volume:
          type: number
          format: double
    HistoricalResponse:
      type: object
      required: [symbol, interval, bars]
      properties:
        symbol:
          type: string
        interval:
          $ref: "#/components/schemas/Timeframe"
        bars:
          type: array
          items:
            $ref: "#/components/schemas/Bar"
        truncated:
          type: boolean
          description: True if the requested range contained more bars than the cap (5000); the most recent are returned.
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable error message.

    # FastAPI emits these for 422 validation failures; mirrored here so the
    # drift test (FastAPI.openapi() == this file) is a plain dict equality.
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
    ValidationError:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string
security:
  - bearer: []